feat: Исправлены замечания ревьюера

+++ main/src/main/java/ru/practicum/user/UserService.java
@@ -6,7 +6,7 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import ru.practicum.exception.UserNotFoundException;
+import ru.practicum.exception.EntityNotFoundException;
 import ru.practicum.user.dto.NewUserRequest;
 import ru.practicum.user.dto.UserDto;
 
@@ -28,12 +28,16 @@ public class UserService {
 
     @Transactional
     public void delete(Long userId) {
-        userRepository.findById(userId)
-                .orElseThrow(() -> new UserNotFoundException("User with id = " + userId + " was not found."));
+        if (!userRepository.existsById(userId)) {
+            throw new EntityNotFoundException("User with id = " + userId + " was not found.");
+        }
         userRepository.deleteById(userId);
     }
 
     public List<UserDto> get(List<Long> ids, Integer from, Integer size) {
+        if (ids == null) {
+            ids = List.of();
+        }
         Pageable pageable = PageRequest.of(from / size, size, Sort.by(ASC, "id"));
         List<User> users = userRepository.findByIdIn(ids, pageable);
         return UserMapper.mapUsersToUserDtos(users);
